fix(observability): address code review findings on telemetry consolidation
- Fix the MERIDIAN_TELEMETRY_DISABLED/MERIDIAN_TRACING_DISABLED kill-switch
  name mismatch between Rust and Python — both sides now read the same var,
  so it actually disables capture everywhere, not just half the pipeline.
- Detect (and warn on) the copytruncate TOCTOU race instead of silently
  losing in-flight log bytes.
- Fix a bash 3.2 "unbound variable" crash in `meridian logs`/`meridian logs
  all` (the empty-array case, also the most common invocation) under macOS's
  stock /bin/bash with `set -u`.
- Restore severity filtering for `*-error` log targets via a new
  `--min-severity` flag on the spool decoder, instead of silently aliasing
  them to their non-error counterpart.
- Fix `is_canonical_install()` to check the running executable's own path
  rather than a machine-wide `~/.meridian/.env` marker, so it no longer
  false-positives (and silently disables OTLP shipping) on a dev machine
  that has ever also had the packaged app installed.
- Fix `meridian logs -f` dropping records that share the exact same
  timestamp as the last-printed record across a poll-tick boundary.
- Document that OTLP-shipping settings fields are now settings.json-only
  (no UI) since AdvancedSection.tsx dropped the config panel.
- Split observability.rs and shipper.rs (both over/further over the
  500-line cap) into smaller focused modules: install_mode, otlp_target,
  launchd_log_cap, retention.
