+++ .env.example
@@ -181,8 +181,9 @@
 # set both to the same value when Jira is configured).
 # JIRA_URL=https://your-org.atlassian.net
 
-# Disable Python OTLP tracing regardless of OO config (1 = off).
-# MERIDIAN_TRACING_DISABLED=0
+# Hard kill switch for local OTel capture (Rust daemon + Python agents both
+# read this same variable — see CLAUDE.md's observability env var table).
+# MERIDIAN_TELEMETRY_DISABLED=0
 
 # Confident AI key for the classifier eval harness (services/tests/evals only).
 # CONFIDENT_API_KEY=
