+++ src/observability/mod.rs
@@ -38,7 +38,7 @@
 // of shipping configuration. Only *delivery* (the background shipper POSTing
 // spooled files to OpenObserve) is gated, and only for a Dev/Bare install with
 // `otlp_enabled` + credentials; a Canonical (packaged/shipped) install never
-// attempts network delivery. See `is_canonical_install()` below and
+// attempts network delivery. See `install_mode::is_canonical_install()` and
 // `telemetry_spool::shipper`.
 //
 // Environment variables read at init time:
@@ -75,31 +75,11 @@ use std::path::PathBuf;
 use std::sync::OnceLock;
 use tracing_subscriber::{layer::SubscriberExt, reload, util::SubscriberInitExt, EnvFilter};
 
-const DEFAULT_OTLP_ENDPOINT: &str = "http://localhost:5080/api/default/v1/traces";
-
-/// True when running from a canonical packaged install (`~/.meridian/.env`
-/// exists — written by the DMG/npm installer; see CLAUDE.md's "Daemon config
-/// gotcha"). A packaged install must never attempt live delivery to
-/// OpenObserve — telemetry capture stays fully local, and the only path to a
-/// developer's OpenObserve is a user-initiated export bundle imported by hand.
-/// A Dev/Bare checkout (no canonical `.env`) may still ship live if
-/// `otlp_enabled` + credentials are configured, for engineers debugging
-/// against their own instance. Mirrors (independently — the daemon has no
-/// dependency on the tray crate) `tray/src-tauri/src/install.rs`'s
-/// `InstallMode::Canonical` check.
-fn is_canonical_install() -> bool {
-    std::env::var("HOME")
-        .map(|home| std::path::Path::new(&home).join(".meridian/.env").exists())
-        .unwrap_or(false)
-}
-
-/// Hard kill switch for OTel capture (spans/logs to the local spool). Plain
-/// stdout/stderr/JSONL-file logs are unaffected. Mirrors Python's
-/// `MERIDIAN_TRACING_DISABLED`.
-fn capture_disabled() -> bool {
-    std::env::var("MERIDIAN_TELEMETRY_DISABLED")
-        .is_ok_and(|v| v == "1" || v.eq_ignore_ascii_case("true"))
-}
+mod install_mode;
+mod otlp_target;
+use install_mode::capture_disabled;
+use otlp_target::DEFAULT_OTLP_ENDPOINT;
+pub use otlp_target::{is_otlp_configured, resolve_otlp_endpoint, resolve_otlp_target, OtlpTarget};
 
 /// Type alias for the hot-reload handle. The `S = Registry` parameter reflects
 /// that the reload layer is installed directly on `tracing_subscriber::Registry`
@@ -246,124 +226,6 @@ pub fn init(service_name: &str) -> Result<ObservabilityGuard> {
         logger_provider,
         otel_enabled,
     })
-}
-
-/// Resolved OTLP export target: trace endpoint + Basic-auth credential.
-/// `None` means export is disabled (toggle off, or no credentials anywhere).
-pub struct OtlpTarget {
-    pub endpoint: String,
-    pub auth: String,
-}
-
-/// Cheap liveness check used by the health probe — does NOT assemble
-/// credentials. Returns `true` when OTLP export would be attempted if
-/// `resolve_otlp_target()` were called (toggle on + credentials present +
-/// not a Canonical/packaged install).
-pub fn is_otlp_configured() -> bool {
-    if is_canonical_install() {
-        return false;
-    }
-    let settings = crate::config::load_runtime_settings();
-    if !settings.otlp_enabled {
-        return false;
-    }
-    // settings.json is the single source for OO credentials — the old
-    // MERIDIAN_OO_AUTH env fallback is deprecated and ignored.
-    settings.oo_email.as_deref().is_some_and(|e| !e.is_empty())
-        && settings
-            .oo_password
-            .as_deref()
-            .is_some_and(|p| !p.is_empty())
-}
-
-/// Resolve the configured OTLP endpoint URL (without assembling credentials).
-/// Used by the health check to derive the `/healthz` URL to ping.
-pub fn resolve_otlp_endpoint() -> Option<String> {
-    if is_canonical_install() {
-        return None;
-    }
-    let settings = crate::config::load_runtime_settings();
-    if !settings.otlp_enabled {
-        return None;
-    }
-    Some(
-        settings
-            .otlp_endpoint
-            .filter(|s| !s.is_empty())
-            .or_else(|| {
-                std::env::var("MERIDIAN_OTLP_ENDPOINT")
-                    .ok()
-                    .filter(|s| !s.is_empty())
-            })
-            .unwrap_or_else(|| DEFAULT_OTLP_ENDPOINT.to_string()),
-    )
-}
-
-/// Resolve the full OTLP export target: endpoint + Basic-auth header value.
-/// Called only at daemon startup (inside `try_build_otel_providers`). Use
-/// `is_otlp_configured()` + `resolve_otlp_endpoint()` for lighter call sites.
-pub fn resolve_otlp_target() -> Option<OtlpTarget> {
-    use base64::{engine::general_purpose::STANDARD, Engine as _};
-
-    // Packaged installs never ship live to OpenObserve — capture stays fully
-    // local; the only path out is a user-initiated export bundle. This check
-    // comes first and unconditionally, regardless of settings.json.
-    if is_canonical_install() {
-        return None;
-    }
-
-    let settings = crate::config::load_runtime_settings();
-
-    if !settings.otlp_enabled {
-        return None;
-    }
-
-    // Auth: settings email+password only. The MERIDIAN_OO_AUTH env fallback is
-    // DEPRECATED and ignored — a dual credential store (env + settings) meant
-    // the UI could show creds while the daemon used different ones (or none).
-    // Credentials are set in the dashboard Settings and read from settings.json.
-    let auth = match (&settings.oo_email, &settings.oo_password) {
-        (Some(email), Some(pass)) if !email.is_empty() && !pass.is_empty() => {
-            // Guard against HTTP header injection and malformed user:password splits.
-            if email.contains(['\n', '\r']) || pass.contains(['\n', '\r']) || email.contains(':') {
-                tracing::warn!(
-                    "OTLP credentials contain invalid characters — OTLP export disabled"
-                );
-                return None;
-            }
-            STANDARD.encode(format!("{email}:{pass}"))
-        }
-        _ => {
-            if std::env::var("MERIDIAN_OO_AUTH").is_ok_and(|v| !v.is_empty()) {
-                tracing::warn!(
-                    "MERIDIAN_OO_AUTH is set but deprecated and ignored — \
-                     set OpenObserve credentials in the dashboard Settings instead"
-                );
-            }
-            return None;
-        }
-    };
-
-    let endpoint = settings
-        .otlp_endpoint
-        .filter(|s| !s.is_empty())
-        .or_else(|| {
-            std::env::var("MERIDIAN_OTLP_ENDPOINT")
-                .ok()
-                .filter(|s| !s.is_empty())
-        })
-        .unwrap_or_else(|| DEFAULT_OTLP_ENDPOINT.to_string());
-
-    // Validate scheme — only http/https are valid OTLP transports.
-    if !endpoint.starts_with("http://") && !endpoint.starts_with("https://") {
-        tracing::warn!(
-            endpoint = %endpoint,
-            "OTLP endpoint has no http/https scheme — OTLP export disabled"
-        );
-        return None;
-    }
-
-    Some(OtlpTarget { endpoint, auth })
 }
 
 /// Builds the OTel tracer and logger providers.
