+++ src/telemetry_spool/retention.rs
@@ -1,0 +1,322 @@
+//ambient dev tool that watches what you do and updates your PM tickets automatically, boosting developer productivity
+//! Local spool housekeeping: crash-orphan sweep, age-based retention, oldest-
+//! first pending-size cap, and the ordering `shipper.rs` ships files in. Split
+//! out of `shipper.rs` (which was over the repo's 500-line cap) since these
+//! are self-contained disk-maintenance operations, independent of the
+//! HTTP-delivery logic the rest of that module owns — they run even on a
+//! Canonical/packaged install that never ships anything.
+//!
+//! # Who calls this
+//! `shipper::run_tick()`, every `MERIDIAN_TELEMETRY_SHIP_INTERVAL_S` (default
+//! 30s), before the ship-target check: orphan sweep → age-based prune (both
+//! `pending/` and `sent/`) → pending-size cap → (only then) ship attempt.
+//!
+//! # Related
+//! - `writer.rs` — the `<signal>-<unix_micros>-<seq>.otlp` filename scheme
+//!   [`list_pending_oldest_first`] and [`enforce_pending_cap`] parse.
+//! - `shipper::run_tick` — the sole caller, and the thing whose delivery
+//!   order [`list_pending_oldest_first`] determines.
+
+use std::{
+    path::{Path, PathBuf},
+    time::{SystemTime, UNIX_EPOCH},
+};
+
+use anyhow::Result;
+
+use crate::telemetry_spool::writer::{micros_from_filename, seq_from_filename};
+
+const DEFAULT_MAX_PENDING_MB: u64 = 512;
+/// `.otlp.tmp` files older than this are crash orphans — a healthy write turns
+/// tmp → final in milliseconds, so anything this old will never be completed.
+const TMP_ORPHAN_MAX_AGE_SECS: u64 = 300;
+
+pub(super) fn max_pending_bytes() -> u64 {
+    let mb = std::env::var("MERIDIAN_TELEMETRY_MAX_PENDING_MB")
+        .ok()
+        .and_then(|v| v.parse::<u64>().ok())
+        .unwrap_or(DEFAULT_MAX_PENDING_MB);
+    mb * 1024 * 1024
+}
+
+/// Remove `.otlp.tmp` files left behind by a crash between write and rename.
+/// A healthy write completes the rename in milliseconds, so any tmp older than
+/// `TMP_ORPHAN_MAX_AGE_SECS` is dead weight the cap/lister would never see.
+pub(super) fn sweep_tmp_orphans(pending: &Path) {
+    let now = SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .unwrap_or_default()
+        .as_secs();
+    let Ok(entries) = std::fs::read_dir(pending) else {
+        return;
+    };
+    for entry in entries.flatten() {
+        let path = entry.path();
+        let Some(name) = path.file_name().and_then(|n| n.to_str()) else {
+            continue;
+        };
+        if !name.ends_with(".otlp.tmp") {
+            continue;
+        }
+        let age = entry
+            .metadata()
+            .and_then(|m| m.modified())
+            .map(|mt| {
+                now.saturating_sub(mt.duration_since(UNIX_EPOCH).unwrap_or_default().as_secs())
+            })
+            .unwrap_or(u64::MAX);
+        if age >= TMP_ORPHAN_MAX_AGE_SECS {
+            let _ = std::fs::remove_file(&path);
+            tracing::debug!(file = %path.display(), age_secs = age, "swept crash-orphaned spool tmp file");
+        }
+    }
+}
+
+/// List `.otlp` files in `dir` sorted oldest-first by `(micros, seq)`.
+///
+/// Including `seq` makes ordering deterministic when two files share a
+/// microsecond (traces+logs back-to-back, or a burst). Files with an
+/// unparseable name are SKIPPED rather than collapsed to key `0` — a renamed /
+/// foreign file must not sort permanently to the front and ship every tick.
+pub(super) fn list_pending_oldest_first(dir: &Path) -> Vec<PathBuf> {
+    let Ok(entries) = std::fs::read_dir(dir) else {
+        return vec![];
+    };
+
+    let mut files: Vec<(u64, u64, PathBuf)> = entries
+        .flatten()
+        .filter_map(|e| {
+            let p = e.path();
+            let name = p.file_name()?.to_str()?.to_string();
+            if !name.ends_with(".otlp") {
+                return None;
+            }
+            let micros = micros_from_filename(&name)?;
+            let seq = seq_from_filename(&name)?;
+            Some((micros, seq, p))
+        })
+        .collect();
+
+    files.sort_by_key(|(m, s, _)| (*m, *s));
+    files.into_iter().map(|(_, _, p)| p).collect()
+}
+
+/// Delete `.otlp` files in `dir` whose mtime is older than `cutoff_secs`.
+/// Used for both `pending/` and `sent/` — retention applies to spooled
+/// telemetry regardless of whether it was ever shipped (a Canonical/packaged
+/// install never ships at all, so `pending/` is its only spool location).
+pub(super) fn prune_by_age(dir: &Path, cutoff_secs: u64) -> Result<()> {
+    let now = SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .unwrap_or_default()
+        .as_secs();
+
+    let Ok(entries) = std::fs::read_dir(dir) else {
+        return Ok(());
+    };
+
+    for entry in entries.flatten() {
+        let path = entry.path();
+        if path.extension().is_none_or(|e| e != "otlp") {
+            continue;
+        }
+        if let Ok(meta) = std::fs::metadata(&path) {
+            if let Ok(mtime) = meta.modified() {
+                let age_secs = now.saturating_sub(
+                    mtime
+                        .duration_since(UNIX_EPOCH)
+                        .unwrap_or_default()
+                        .as_secs(),
+                );
+                // `>=` so a retention of 0 days means "keep nothing" (prune even
+                // just-shipped files); the default 7-day cutoff is unaffected since
+                // a fresh file's age (~0s) is never >= 604800s.
+                if age_secs >= cutoff_secs {
+                    let _ = std::fs::remove_file(&path);
+                    tracing::debug!(file = %path.display(), age_days = age_secs / 86400, "pruned old telemetry file");
+                }
+            }
+        }
+    }
+    Ok(())
+}
+
+/// Drop OLDEST pending files beyond the size cap with a structured warning.
+/// Never silently drops — always emits `tracing::warn!` with count + bytes.
+pub(super) fn enforce_pending_cap(pending: &Path) -> Result<()> {
+    let max = max_pending_bytes();
+
+    let Ok(entries) = std::fs::read_dir(pending) else {
+        return Ok(());
+    };
+
+    let mut files: Vec<(u64, u64, u64, PathBuf)> = entries // (micros, seq, size, path)
+        .flatten()
+        .filter_map(|e| {
+            let p = e.path();
+            let name = p.file_name()?.to_str()?.to_string();
+            if !name.ends_with(".otlp") {
+                return None;
+            }
+            let micros = micros_from_filename(&name)?;
+            let seq = seq_from_filename(&name)?;
+            let size = std::fs::metadata(&p).map(|m| m.len()).unwrap_or(0);
+            Some((micros, seq, size, p))
+        })
+        .collect();
+
+    // Sort oldest-first by (micros, seq) — same total order the shipper uses, so
+    // the cap evicts the genuinely-oldest records, not whatever read_dir yields.
+    files.sort_by_key(|(m, s, _, _)| (*m, *s));
+
+    let total: u64 = files.iter().map(|(_, _, s, _)| s).sum();
+    if total <= max {
+        return Ok(());
+    }
+
+    let mut to_drop = total - max;
+    let mut dropped_count = 0u64;
+    let mut dropped_bytes = 0u64;
+
+    for (_, _, size, path) in &files {
+        if to_drop == 0 {
+            break;
+        }
+        let _ = std::fs::remove_file(path);
+        dropped_bytes += size;
+        dropped_count += 1;
+        to_drop = to_drop.saturating_sub(*size);
+    }
+
+    if dropped_count > 0 {
+        tracing::warn!(
+            dropped_files = dropped_count,
+            dropped_bytes,
+            cap_bytes = max,
+            "pending telemetry cap exceeded — oldest spool files dropped"
+        );
+    }
+
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::telemetry_spool::writer::{pending_dir, write_pending};
+    use tempfile::TempDir;
+
+    #[test]
+    fn list_pending_sorted_oldest_first() {
+        let dir = TempDir::new().unwrap();
+        let base = dir.path().to_path_buf();
+
+        // Write files with different seq numbers; filename micros will be equal
+        // or ascending since they run in sequence. Seq counter disambiguates.
+        let p1 = write_pending(&base, "traces", b"old").unwrap();
+        let p2 = write_pending(&base, "traces", b"new").unwrap();
+
+        let pending = pending_dir(&base);
+        let sorted = list_pending_oldest_first(&pending);
+        assert_eq!(sorted.len(), 2);
+        // First file written should appear first (lower seq or lower micros)
+        assert_eq!(sorted[0], p1);
+        assert_eq!(sorted[1], p2);
+    }
+
+    #[test]
+    fn list_pending_orders_same_micros_by_seq_and_skips_unparseable() {
+        let dir = TempDir::new().unwrap();
+        let pending = pending_dir(dir.path());
+        std::fs::create_dir_all(&pending).unwrap();
+
+        // Same microsecond, out-of-order seq — must come back seq 0 then seq 1.
+        std::fs::write(pending.join("traces-1000-1.otlp"), b"b").unwrap();
+        std::fs::write(pending.join("traces-1000-0.otlp"), b"a").unwrap();
+        // A crash-orphan tmp and a foreign name must be ignored entirely (the
+        // old `unwrap_or(0)` would have sorted the foreign file permanently first).
+        std::fs::write(pending.join("traces-1000-2.otlp.tmp"), b"x").unwrap();
+        std::fs::write(pending.join("garbage.otlp"), b"y").unwrap();
+
+        let sorted = list_pending_oldest_first(&pending);
+        assert_eq!(sorted.len(), 2, "tmp + foreign names excluded");
+        assert!(sorted[0]
+            .file_name()
+            .unwrap()
+            .to_str()
+            .unwrap()
+            .ends_with("-0.otlp"));
+        assert!(sorted[1]
+            .file_name()
+            .unwrap()
+            .to_str()
+            .unwrap()
+            .ends_with("-1.otlp"));
+    }
+
+    #[test]
+    fn prune_by_age_removes_old_files_in_sent() {
+        let dir = TempDir::new().unwrap();
+        let sent = dir.path().join("sent");
+        std::fs::create_dir_all(&sent).unwrap();
+
+        let old_file = sent.join("traces-1-0.otlp");
+        std::fs::write(&old_file, b"x").unwrap();
+        // A 0-second cutoff means "keep nothing" — everything is at least 0s old.
+        prune_by_age(&sent, 0).unwrap();
+
+        assert!(!old_file.exists());
+    }
+
+    #[test]
+    fn prune_by_age_removes_old_files_in_pending() {
+        // Retention must apply to pending/ too — it's the ONLY spool location
+        // for a Canonical/packaged install, which never ships to sent/.
+        let dir = TempDir::new().unwrap();
+        let base = dir.path().to_path_buf();
+        let p = write_pending(&base, "traces", b"x").unwrap();
+        let pending = pending_dir(&base);
+
+        prune_by_age(&pending, 0).unwrap();
+
+        assert!(!p.exists());
+    }
+
+    #[test]
+    fn prune_by_age_keeps_fresh_files() {
+        let dir = TempDir::new().unwrap();
+        let base = dir.path().to_path_buf();
+        let p = write_pending(&base, "traces", b"x").unwrap();
+        let pending = pending_dir(&base);
+
+        // A generous cutoff (1 day) must not touch a file written moments ago.
+        prune_by_age(&pending, 24 * 3600).unwrap();
+
+        assert!(p.exists());
+    }
+
+    #[test]
+    fn pending_cap_drops_oldest_first_with_warn() {
+        // Set cap to 1 byte so everything over 1 byte gets dropped
+        std::env::set_var("MERIDIAN_TELEMETRY_MAX_PENDING_MB", "0");
+
+        let dir = TempDir::new().unwrap();
+        let base = dir.path().to_path_buf();
+        write_pending(&base, "traces", b"aaa").unwrap();
+        write_pending(&base, "traces", b"bbb").unwrap();
+
+        let pending = pending_dir(&base);
+        // Cap of 0 MB → all files should be dropped
+        enforce_pending_cap(&pending).unwrap();
+
+        let remaining: Vec<_> = std::fs::read_dir(&pending)
+            .unwrap()
+            .flatten()
+            .filter(|e| e.path().extension().is_some_and(|x| x == "otlp"))
+            .collect();
+        // With 0MB cap all are dropped
+        assert!(remaining.is_empty());
+
+        std::env::remove_var("MERIDIAN_TELEMETRY_MAX_PENDING_MB");
+    }
+}
